Break out drawer and change logo to reg text
