+++ src/App.vue
@@ -1,45 +1,13 @@
 <script setup lang="ts">
 import Header from "@Components/Body/Header.vue";
+import Drawer from "@Components/Body/Drawer.vue";
 import { useAppState } from "./State/AppState";
 
 const app$ = useAppState();
-
-const getDrawerClass = () => {
-  console.log(app$.Layout.DrawerIsOpen ? 'drawer-open' : 'drawer-shut');
-  return app$.Layout.DrawerIsOpen ? 'drawer-open' : 'drawer-shut';
-}
 </script>
 
 <template>
-  <!--==== Drawer =======================================================-->
-  <div 
-    class="drawer"
-    :class="getDrawerClass()"
-  >
-
-    <div class="drawer-header">
-      <button 
-        class="flex-button"
-        @click="app$.Layout.ToggleDrawer()"
-        >
-          👈
-          <img 
-            src="../public/branding/logo-wide.png"
-            style="height:100%"
-          /> 
-      </button>
-    </div>
-
-    <div class="drawer-content">
-      <button class="drawer-button"><span>🫵 One</span></button>
-      <button class="drawer-button"><span>🫡 Two</span></button>
-      <button class="drawer-button"><span>🎼 Three</span></button>
-      <button class="drawer-button"><span>🎠 Four</span></button>
-    </div>
-
-  </div>
-
-  <!--==== App-Container ================================================-->
+  <drawer v-show="app$.Layout.DrawerIsOpen"></drawer>
   <div id="root" class="app-container">
 
     <!--==== Header Ad ==================================================-->
@@ -61,9 +29,7 @@ const getDrawerClass = () => {
     <!--==== Content ====================================================-->
 
     <div class="content-wrapper debug-border">
-
       <div>Content</div>
-
     </div>
     
     <!--==== ======= ====================================================-->
@@ -90,10 +56,6 @@ const getDrawerClass = () => {
 
 <style scoped lang="scss">
 
-// .debug-border{
-//   border:1px dotted grey;
-// }
-
 .app-container {
   height: 100vh;
   width:100%;
@@ -114,90 +76,6 @@ const getDrawerClass = () => {
   @apply w-full flex justify-center;
 
 
-}
-
-.drawer{
-  position: absolute;
-  z-index: 10;
-  overflow: hidden;
-  top:0;
-  bottom:0;
-  max-width:250px;
-  display: grid;
-  grid-template-rows: 0em 2.5em 1fr 2.5em 0em;
-  
-
-  transition: 150ms;
-  transition-timing-function: ease-out;
-
-}
-.drawer-open{
-  width:100%;
-  background-color: black;
-  border-right: 4px double grey;
-
-}
-.drawer-shut{
-  width:0%;
-  border-right: none;
-
-}
-
-
-
-.drawer-header{  
-  grid-row:2;
-  display: flex;
-  margin-left: 1em;
-}
-
-.drawer-content{
-  grid-row:3;
-  max-width:250px;
-  
-}
-
-.drawer-button{
-  font-size:x-large;
-  background-color: transparent;
-  border: none;
-  color: white;
-  width:100%
-  
-}
-.drawer-button:hover{
-  color: goldenrod;
-  
-}
-.drawer-button:active{
-  color: gold;
-}
-
-.flex-button{
-  display: flex;
-  height:100%;  
-  width:100%;
-  max-width:220px;
-  font-size:xx-large;
-  background-color: transparent;
-  border: none;
-  
-  align-items:flex-end;
-  margin-left:.75em;
-
-  border-radius: 50%;
-
-}
-.flex-button:hover{
-  box-shadow:
-    -14px -5px 5px rgb(65, 46, 0),
-    14px 5px 5px goldenrod; 
-  
-}
-.flex-button:active{
-  box-shadow:
-    -14px -5px 5px rgb(53, 14, 0),
-    14px 5px 5px rgb(86, 23, 0),
 }
 
 </style>
