+++ src/Config/Styles/primevue.css
@@ -1,4 +1,4 @@
-.p-component, .p-component * {
+/* .p-component, .p-component * {
     box-sizing: border-box;
 }
 
@@ -490,4 +490,4 @@ button {
     left: 50%;
     margin-left: -.25rem;
     border-width: 0 .25em .25rem;
-}
+} */
