+++ src/ViewComponents/Body/LogoButton.vue
@@ -1,0 +1,32 @@
+<script setup lang="ts">
+
+</script>
+
+<!-- === -->
+<template>
+  <div class="logo-button">
+    <span>Shogi&nbsp;Time</span>
+    <img 
+      src="../../../public/branding/logo-icon.png"
+    /> 
+  </div>
+</template>
+
+<!-- === -->
+<style>
+.logo-button{
+  white-space: nowrap;
+  
+}
+.logo-button span{
+  font-size: xx-large;
+  vertical-align:baseline;
+  font-variant: small-caps;
+  
+}
+
+.logo-button img{
+  height:2em;
+}
+
+</style>
